Refactor logging. Add validation for path variables

+++ src/main/java/com/godel/simplecrud/controller/EmployeeController.java
@@ -1,11 +1,13 @@
 package com.godel.simplecrud.controller;
 
+import com.godel.simplecrud.exceptions.EmployeeControllerIllegalArgumentException;
 import com.godel.simplecrud.model.Employee;
 import com.godel.simplecrud.service.EmployeeService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -33,14 +35,22 @@ public class EmployeeController {
     @GetMapping("/{id}")
     @Operation(summary = "Show specific employee by his ID", description = "Let see specific employee by his ID as path variable")
     public Employee showEmployee(@PathVariable @Parameter(description = "Employee ID") Long id) {
+        if (id <= 0) {
+            throw new EmployeeControllerIllegalArgumentException("ID cannot be less or equal 0");
+        }
+
         return employeeService.findEmployeeById(id);
     }
 
     @GetMapping("{first_name}/{last_name}")
     public Employee showEmployeeByFirstNameAndLastName(@PathVariable String first_name,
                                                        @PathVariable String last_name) {
-        //todo validation
-        return employeeService.findByFirstNameAndLastName(first_name, last_name);
+        String pattern = "\\D{3,}";
+        if (!first_name.matches(pattern) || !last_name.matches(pattern)) {
+            throw new EmployeeControllerIllegalArgumentException("First name and last name must have length 3+ letters and contains no numbers");
+        }
+
+        return employeeService.findByFirstNameAndLastName(StringUtils.capitalize(first_name), StringUtils.capitalize(last_name));
     }
 
     @PostMapping
@@ -52,8 +62,10 @@ public class EmployeeController {
 
     @PutMapping("/{id}")
     @Operation(summary = "Create or update employee", description = "Let create new employee with given parameters or update if he already exists")
-    public Employee updateEmployee(@PathVariable Long id,
-                                                   @RequestBody Employee employee) {
+    public Employee updateEmployee(@PathVariable Long id, @RequestBody Employee employee) {
+        if (id <= 0) {
+            throw new EmployeeControllerIllegalArgumentException("ID cannot be less or equal 0");
+        }
         employee.setEmployeeId(id);
 
         return employeeService.updateOrCreateEmployee(employee);
@@ -62,6 +74,10 @@ public class EmployeeController {
     @DeleteMapping("/{id}")
     @Operation(summary = "Delete employee", description = "Let delete employee by his ID")
     public void deleteEmployee(@PathVariable @Parameter(description = "Employee ID") Long id) {
+        if (id <= 0) {
+            throw new EmployeeControllerIllegalArgumentException("ID cannot be less or equal 0");
+        }
+
         employeeService.deleteEmployeeById(id);
     }
 }
